+++ lib/skogsra/docs.ex
@@ -40,44 +40,19 @@ defmodule Skogsra.Docs do
 
   defp do_gen_full_docs(module, function_name, app_name, keys, options, docs) do
     module = Macro.to_string(module)
-    no_namespace = Env.new(nil, app_name, keys, options)
-    namespace = My.Custom.Namespace
-    with_namespace = Env.new(namespace, app_name, keys, options)
+    env = Env.new(nil, app_name, keys, options)
 
     """
     #{insert_custom_docs(docs)}
 
-    Calling `#{module}.#{function_name}()` will try to retrive a casted value
-    in the following order:
+    Calling `#{module}.#{function_name}()` will ensure the following:
 
-    1. From the OS environment variable
-       `#{inspect(Env.os_env(no_namespace))}` (unless `skip_system: true`).
-    2. From the application config (unless `skip_config: true`) e.g:
-       ```
-       #{gen_config_code(no_namespace)}
-       ```
-    3. From the default value if defined. If the variables has the option
-       `required: true` and no default is defined, then the function will
-       return an error.
-
-    Calling `#{module}.#{function_name}(#{inspect(namespace)})` will try to
-    retrieve a casted value in the following order:
-
-    1. From the OS environment variable
-       `#{inspect(Env.os_env(with_namespace))}` (unless `skip_system: true`).
-    2. From the application config (unless `skip_config: true`) e.g:
-       ```
-       #{gen_config_code(with_namespace)}
-       ```
-    3. From the OS environment variable
-       `#{inspect(Env.os_env(no_namespace))}` (unless `skip_system: true`).
-    4. From the application config (unless `skip_config: true`) e.g:
-       ```
-       #{gen_config_code(no_namespace)}
-       ```
-    5. From the default value if defined. If the variables has the option
-       `required: true` and no default is defined, then the function will
-       return an error.
+    - Binding order: #{inspect(Env.binding_order(env))}
+    - OS environment variable: #{inspect(Env.os_env(env))}
+    - Type: #{inspect(Env.type(env))}
+    - Default: #{inspect(Env.default(env))}
+    - Required: #{inspect(Env.required?(env))}
+    - Cached: #{inspect(Env.cached?(env))}
     """
   end