+++ lib/skogsra/env.ex
@@ -25,6 +25,16 @@ defmodule Skogsra.Env do
   @type keys :: [key()]
 
   @typedoc """
+  Variable binding.
+  """
+  @type binding :: :config | :system | module()
+
+  @typedoc """
+  Variable binding list.
+  """
+  @type bindings :: [binding()]
+
+  @typedoc """
   Types.
   """
   @type type ::
@@ -39,8 +49,8 @@ defmodule Skogsra.Env do
 
   @typedoc """
   Environment variable options.
-  - `skip_system` - Skips loading the variable from the OS.
-  - `skip_config` - Skips loading the variable from the config.
+  - `load_order` - Variable binding load order.
+  - `skip` - Skips loading a variable from the list of bindings.
   - `os_env` - The name of the OS environment variable.
   - `type` - Type to cast the OS environment variable value.
   - `namespace` - Default namespace for the variable.
@@ -49,8 +59,8 @@ defmodule Skogsra.Env do
   - `cached` - Whether the variable is cached or not.
   """
   @type option ::
-          {:skip_system, boolean()}
-          | {:skip_config, boolean()}
+          {:load_order, bindings()}
+          | {:skip, bindings()}
           | {:os_env, binary()}
           | {:type, type()}
           | {:namespace, namespace()}
@@ -105,33 +115,11 @@ defmodule Skogsra.Env do
   end
 
   @doc """
-  Whether the `Skogsra` environment variable should skip system or not.
-  """
-  @spec skip_system?(t()) :: boolean()
-  def skip_system?(%Env{options: options}) do
-    case options[:skip_system] do
-      true -> true
-      _ -> false
-    end
-  end
-
-  @doc """
-  Whether the `Skogsra` environment variable should skip config or not.
-  """
-  @spec skip_config?(t()) :: boolean()
-  def skip_config?(%Env{options: options}) do
-    case options[:skip_config] do
-      true -> true
-      _ -> false
-    end
-  end
-
-  @doc """
   Gets the OS variable name for the `Skogsra` environment variable.
   """
   @spec os_env(t()) :: binary()
   def os_env(%Env{options: options} = env) do
-    with false <- Env.skip_system?(env),
+    with true <- :system in Env.binding_order(env),
          value when not is_binary(value) <- options[:os_env] do
       namespace = gen_namespace(env)
       app_name = gen_app_name(env)
@@ -139,7 +127,7 @@ defmodule Skogsra.Env do
 
       "#{namespace}#{app_name}_#{keys}"
     else
-      true -> ""
+      false -> ""
       value -> value
     end
   end
@@ -186,16 +174,53 @@ defmodule Skogsra.Env do
     end
   end
 
+  @doc """
+  Gets the binding order for a `Skogsra` environment variable.
+  """
+  @spec binding_order(t()) :: bindings()
+  def binding_order(%Env{options: options}) do
+    options[:binding_order] -- options[:binding_skip]
+  end
+
   #########
   # Helpers
 
   @doc false
+  @spec defaults(options()) :: options()
   def defaults(options) do
     options
-    |> Keyword.put_new(:skip_system, false)
-    |> Keyword.put_new(:skip_config, false)
     |> Keyword.put_new(:required, false)
     |> Keyword.put_new(:cached, true)
+    |> Keyword.put_new(:binding_order, binding_order())
+    |> Keyword.put_new(:binding_skip, binding_skip())
+  end
+
+  @doc false
+  @spec binding_order() :: bindings()
+  def binding_order do
+    default = [:system, :config]
+    bindings = Application.get_env(:skogsra, :binding_order)
+
+    if is_bindings?(bindings), do: bindings, else: default
+  end
+
+  @doc false
+  @spec binding_skip() :: bindings()
+  def binding_skip do
+    default = []
+    bindings = Application.get_env(:skogsra, :binding_skip)
+
+    if is_bindings?(bindings), do: bindings, else: default
+  end
+
+  @doc false
+  @spec is_bindings?(term()) :: boolean()
+  def is_bindings?(other) when not is_list(other), do: false
+
+  def is_bindings?(bindings) do
+    Enum.all?(bindings, fn binding ->
+      binding in [:system, :config] or Code.ensure_loaded?(binding)
+    end)
   end
 
   @doc false