+++ test/skogsra/sys_test.exs
@@ -5,32 +5,21 @@ defmodule Skogsra.SysTest do
   alias Skogsra.Sys
 
   describe "get_env/1" do
-    test "when skip_system is true, skips it" do
-      env = Env.new(nil, :system_app, [:a, :b], skips_system: true)
+    setup do
+      name = "VAR#{make_ref() |> :erlang.phash2()}"
+      env = Env.new(nil, :system_app, :key, os_env: name)
 
-      assert nil == Sys.get_env(env)
+      {:ok, env: env}
     end
 
-    test "when defined, gets the value" do
-      SystemMock.put_env("SYSTEM_APP_A_B", "42")
-
-      env = Env.new(nil, :system_app, [:a, :b], default: 21)
-
-      assert 42 == Sys.get_env(env)
+    test "when variable is not defined, returns nil", %{env: env} do
+      assert {:ok, nil} == Sys.get_env(env)
     end
 
-    test "when undefined, does not get the value" do
-      env = Env.new(nil, :system_app, [:a, :c], [])
+    test "when variable is defined, returns value", %{env: env} do
+      SystemMock.put_env(env.options[:os_env], "42")
 
-      assert nil == Sys.get_env(env)
-    end
-
-    test "when cast is not possible, does not get the value" do
-      SystemMock.put_env("SYSTEM_APP_KEY", "42.0")
-
-      env = Env.new(nil, :system_app, :key, type: :integer)
-
-      assert nil == Sys.get_env(env)
+      assert {:ok, "42"} = Sys.get_env(env)
     end
   end
 end